requirements.txt
discord.py==1.7.3
pymongo==4.0.1
python-dotenv==0.19.2